Explore source of unused function warnings
We add the file src/main.rs, which creates another crate.  We import the
communicator crate, and call its client::connect() function.  This fails
because it is private (by default), which is also the source of the
warnings.  A private function can only be called from within the same
crate, so if it isn't called there, it can never be called, and rustc
warns about unused code.